requirements.txt
# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Pydantic (validation)
pydantic==2.9.2
pydantic-settings==2.5.2

# ASGI server
uvicorn==0.30.6

# Optional: Development tools
python-dotenv==1.0.1

# Optional: Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Optional: Security (production)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
python-binance==1.0.29








# fastapi==0.109.0
# uvicorn[standard]==0.27.0
# sqlalchemy==2.0.25
# pydantic==2.5.3
# python-dotenv==1.0.0
# pydantic-settings==2.1.0

# # Database drivers
# # SQLite (built-in with Python)
# # PostgreSQL için: psycopg2-binary==2.9.9

# # Security
# passlib[bcrypt]==1.7.4

# # HTTP requests (Binance API için)
# httpx==0.26.0
# aiohttp==3.9.1

# # Testing
# pytest==7.4.4
# pytest-asyncio==0.23.3
# httpx==0.26.0